Determine the session automatically for viewer-based credentials.
This commit makes a major overhaul of the API so that we determine the
current Shiny session automatically. It generally makes consuming this
package much easier, and fits the "default credentials" model used by
various other authentication packages more nicely.

The removal of the session parameter also has major implications for
testing, and the mocking story here is now more comprehensive. As a
result, it is no longer necessary to export an "example" Shiny session.

Unit tests and documentation updates are included.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>